refactor: share URL mapping
