+++ popup.html
@@ -123,6 +123,7 @@
     Powered by <a href="http://www.absoluterecruiter.com" target="_blank">Absolute Recruiter</a>
   </div>
 
+  <script src="urls.js"></script>
   <script src="popup.js"></script>
 </body>
 </html>
